+++ fyrox-core/Cargo.toml
@@ -17,6 +17,8 @@ rust-version = "1.80"
 [dependencies]
 fyrox-core-derive = { path = "../fyrox-core-derive", version = "0.36.2" }
 fyrox-math = { path = "../fyrox-math", version = "0.36.2" }
+fyrox-utils = { path = "../fyrox-utils", version = "0.36.2" }
+fyrox-reflect= { path = "../fyrox-reflect", version = "0.36.2" }
 base64 = "0.22.1"
 byteorder = "1.4.3"
 rand = "0.8.4"
@@ -25,12 +27,12 @@ lazy_static = "1.4.0"
 nalgebra = { version = "0.33", features = ["bytemuck"] }
 arrayvec = "0.7.2"
 futures = { version = "0.3.17", features = ["thread-pool"] }
-uuid = { version = "1", features = ["v4", "js"] }
+uuid = { workspace = true, features = ["v4", "js"] }
 instant = { version = "0.1.12", features = ["wasm-bindgen"] }
 num-traits = "0.2.14"
-parking_lot = "0.12.0"
+parking_lot = { workspace = true }
 fxhash = "0.2.1"
-bitflags = "2.2.1"
+bitflags = { workspace = true }
 once_cell = "1.17.1"
 notify = "8"
 serde = { version = "1", features = ["derive"] }
